Revert to fp32 compute precision for model accuracy
fp16 rmsnorm outputs caused fp16 matmuls which, combined with int4
weight quantization, accumulated too much error across 35 layers
producing garbage tokens.

Changes:
- Revert _rmsnorm/_rmsnorm_noscale to return fp32 (all matmuls fp32
  via JAX type promotion — matches origin/main behavior)
- Revert GELU wrapper (fp32 rmsnorm means GELU inputs are already
  fp32, no explicit cast needed for erf polynomial pattern matching)
- Restore softmax .astype(jnp.float32) for numerical safety
- Revert compute_precision to FLOAT32
- Include softmax pattern fix for fp16 cast ops in
  replace_decomposed_softmax.py (cherry-picked)

Verified: JAX and CoreML both produce coherent output. All MIL fusions
intact (35 SDPA, 70 GELU, 35 softmax). Model size 2.84 GB (int4).

+++ gemma_chat/export.py
@@ -276,7 +276,7 @@ def _hlo_to_mlpackage(
             convert_kwargs = dict(
                 source="milinternal",
                 minimum_deployment_target=ct.target.iOS18,
-                compute_precision=ct.precision.FLOAT16,
+                compute_precision=ct.precision.FLOAT32,
                 pass_pipeline=pipeline,
                 skip_model_load=True,
             )
